requirements.txt
absl-py==2.1.0
aggdraw==1.3.16
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Automat==24.8.1
buildtools==1.0.6
causal_conv1d-1.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
cloudpickle==3.1.0
cmake==3.30.4
constantly==23.10.4
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
docopt==0.6.2
einops==0.8.0
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
furl==2.1.3
graphviz==0.20.3
greenlet==3.1.1
h5py==3.11.0
huggingface-hub==0.25.2
hyperlink==21.0.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
incremental==24.7.2
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lit==18.1.8
mamba_ssm-2.2.2
MarkupSafe==2.1.5
matplotlib==3.7.5
MedPy==0.5.2
ml-collections==0.1.1
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
nibabel==5.2.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
orderedmultidict==1.0.1
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
propcache==0.2.0
protobuf==5.28.2
psutil==6.1.0
ptflops==0.7.3
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.2
redo==3.0.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
SimpleITK==2.4.0
simplejson==3.19.3
six==1.16.0
SQLAlchemy==2.0.36
submitit==1.5.2
sympy==1.13.3
tensorboardX==2.6.2.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.20.1
tomli==2.0.2
torch==2.0.0+cu118
torch-geometric==2.6.1
torchaudio==2.0.0+cu118
torchinfo==1.8.0
torchsummary==1.5.1
torchview==0.2.6
torchvision==0.15.1+cu118
tqdm==4.66.5
transformers==4.45.2
triton==2.0.0
Twisted==24.10.0
typing_extensions==4.12.2
urllib3==2.2.3
visualkeras==0.1.3
yacs==0.1.8
yarl==1.15.2
zipp==3.20.2
zope.interface==7.1.1
